+++ cli/root.go
@@ -8,10 +8,18 @@ import (
 	"go.bbkane.com/warg/value/scalar"
 )
 
-func InitCmd() command.Command {
+func InitZshCmd() command.Command {
 	return command.New(
-		"Print script to run envelope when the directory changes. Only zsh supported (for now)",
-		initRun,
+		"Print zsh",
+		initZshRun,
+		command.Flag(
+			"--print-autoload",
+			"Include autoload -Uz add-zsh-hook line (might not be needed if you already autoloaded it)",
+			scalar.Bool(
+				scalar.Default(true),
+			),
+			flag.Required(),
+		),
 		command.Flag(
 			"--print-export-env",
 			"Include export-env/unexport-env to easily use envs from the CLI",
@@ -31,26 +39,37 @@ func InitCmd() command.Command {
 	)
 }
 
-func initRun(cmdCtx command.Context) error {
+func initZshRun(cmdCtx command.Context) error {
 
+	printAutoload := cmdCtx.Flags["--print-autoload"].(bool)
 	printChpwdHook := cmdCtx.Flags["--print-chpwd-hook"].(bool)
 	printExportEnv := cmdCtx.Flags["--print-export-env"].(bool)
 
-	chpwdHook := `
+	prelude := `
 # https://github.com/bbkane/envelope/
 #
 # To initialize envelope, add this to your configuration (usually ~/.zshrc):
 #
 # eval "$(envelope init)"
 #
+`
+	fmt.Fprint(cmdCtx.Stdout, prelude)
+
+	autoload := `
 autoload -Uz add-zsh-hook
+`
+	if printAutoload {
+		fmt.Fprint(cmdCtx.Stdout, autoload)
+	}
+
+	chpwdHook := `
 add-zsh-hook -Uz chpwd (){
     eval $(envelope env print-script --name "$OLDPWD" --no-env-no-problem true --type unexport)
     eval $(envelope env print-script --name "$PWD" --no-env-no-problem true --type export)
 }
 `
 	if printChpwdHook {
-		fmt.Fprintln(cmdCtx.Stdout, chpwdHook)
+		fmt.Fprint(cmdCtx.Stdout, chpwdHook)
 	}
 
 	exportEnv := `
@@ -58,7 +77,7 @@ export-env() { eval $(envelope env print-script --name "$1" --no-env-no-problem 
 unexport-env() { eval $(envelope env print-script --name "$1" --no-env-no-problem true --type unexport) }
 `
 	if printExportEnv {
-		fmt.Fprintln(cmdCtx.Stdout, exportEnv)
+		fmt.Fprint(cmdCtx.Stdout, exportEnv)
 	}
 
 	return nil